handle sizeof(time_t) > sizeof(long) in format strings
Last-Update: 2024-03-11
Forwarded: no

64-bit time_t means that on some architectures, time_t is now larger than
a long, and making some references in format strings incorrect.  To avoid
truncation or other size mismatch issues, always cast to a long long and
read using %lld.

Fixes an assertion failure detected during build-time tests on armhf:
slapd: ../../../../../servers/slapd/overlays/dds.c:422: dds_op_add: Assertion `bv.bv_len < sizeof( ttlbuf )' failed.

Gbp-Pq: Name 64-bit-time-t-compat.patch

+++ servers/slapd/back-asyncmeta/modrdn.c
@@ -256,8 +256,8 @@ asyncmeta_back_modrdn( Operation *op, SlapReply *rs )
 	      op->o_req_dn.bv_val );
 
 	if (current_time > op->o_time) {
-		Debug(asyncmeta_debug, "==> asyncmeta_back_modrdn[%s]: o_time:[%ld], current time: [%ld]\n",
-		      op->o_log_prefix, op->o_time, current_time );
+		Debug(asyncmeta_debug, "==> asyncmeta_back_modrdn[%s]: o_time:[%lld], current time: [%lld]\n",
+		      op->o_log_prefix, (long long)op->o_time, (long long)current_time );
 	}
 	asyncmeta_new_bm_context(op, rs, &bc, mi->mi_ntargets, mi );
 	if (bc == NULL) {
